refactor: move cmd to root and client to app
