+++ app/app.go
@@ -102,7 +102,7 @@ import (
 	ecocreditmodule "github.com/regen-network/regen-ledger/x/ecocredit/module"
 
 	// unnamed import of statik for swagger UI support
-	_ "github.com/regen-network/regen-ledger/v4/client/docs/statik"
+	_ "github.com/regen-network/regen-ledger/v4/app/client/docs/statik"
 )
 
 const (
