+++ app/client/docs/.gitignore
@@ -1,0 +1,5 @@
+regen
+swagger-ui
+swagger.yaml
+swagger-ibc.yaml
+swagger-sdk.yaml
