--- app/regen/main.go
@@ -1,24 +1,0 @@
-package main
-
-import (
-	"os"
-
-	"github.com/cosmos/cosmos-sdk/server"
-	svrcmd "github.com/cosmos/cosmos-sdk/server/cmd"
-
-	"github.com/regen-network/regen-ledger/v4/app"
-	cmd "github.com/regen-network/regen-ledger/v4/app/regen/cmd"
-)
-
-// In main we call the rootCmd
-func main() {
-	rootCmd, _ := cmd.NewRootCmd()
-	if err := svrcmd.Execute(rootCmd, app.EnvPrefix, app.DefaultNodeHome); err != nil {
-		switch e := err.(type) {
-		case server.ErrorCode:
-			os.Exit(e.Code)
-		default:
-			os.Exit(1)
-		}
-	}
-}
