--- client/docs/.gitignore
@@ -1,5 +1,0 @@
-regen
-swagger-ui
-swagger.yaml
-swagger-ibc.yaml
-swagger-sdk.yaml
