--- client/docs/config.json
@@ -1,58 +1,0 @@
-{
-    "swagger": "2.0",
-    "info": {
-        "title": "Regen Ledger - REST and gRPC Gateway Documentation",
-        "description": "Documentation describing REST and gRPC endpoints for state queries and broadcasting transactions.",
-        "version": "v4.0.0"
-    },
-    "apis": [
-        {
-            "url": "./client/docs/regen/data/v1/query.swagger.json"
-        },
-        {
-            "url": "./client/docs/regen/ecocredit/v1/query.swagger.json",
-            "operationIds": {
-                "rename": {
-                  "Balance": "Ecocredit/Balance",
-                  "Params": "Ecocredit/Params"
-                }
-              }
-        },
-        {
-            "url": "./client/docs/regen/ecocredit/basket/v1/query.swagger.json"
-        },
-        {
-            "url": "./client/docs/regen/ecocredit/marketplace/v1/query.swagger.json"
-        },
-        {
-            "url": "./client/docs/regen/group/v1alpha1/query.swagger.json",
-            "operationIds": {
-                "rename": {
-                  "Proposal": "Group/Proposal"
-                }
-              }
-        },
-        {
-            "url": "./client/docs/swagger-sdk.yaml",
-            "dereference": {
-                "circular": "ignore"
-            },
-            "tags": {
-                "rename": {
-                "Gaia REST": "Tendermint RPC"
-                }
-            },
-            "operationIds": {
-                "rename": {
-                  "UpgradedConsensusState": "sdk/UpgradedConsensusState"
-                }
-              }
-        },
-        {
-            "url": "./client/docs/swagger-ibc.yaml",
-            "dereference": {
-                "circular": "ignore"
-            }
-        }
-    ]
-}
