+++ cmd/regen/main.go
@@ -1,0 +1,23 @@
+package main
+
+import (
+	"os"
+
+	"github.com/cosmos/cosmos-sdk/server"
+	"github.com/cosmos/cosmos-sdk/server/cmd"
+
+	"github.com/regen-network/regen-ledger/v4/app"
+	"github.com/regen-network/regen-ledger/v4/app/client/cli"
+)
+
+func main() {
+	rootCmd, _ := cli.NewRootCmd()
+	if err := cmd.Execute(rootCmd, app.EnvPrefix, app.DefaultNodeHome); err != nil {
+		switch e := err.(type) {
+		case server.ErrorCode:
+			os.Exit(e.Code)
+		default:
+			os.Exit(1)
+		}
+	}
+}
